[GILJOB-136] 포지션 퀘스트 별 조회 uri 수정
